refactoring(file): refactoring of the return/throw position
