+++ src/main/java/fr/esgi/grp9/uparserbackend/file/service/IFileService.java
@@ -1,11 +1,13 @@
-package fr.esgi.grp9.uparserbackend.file.domain;
+package fr.esgi.grp9.uparserbackend.file.service;
+
+import fr.esgi.grp9.uparserbackend.file.domain.File;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface FileService {
+public interface IFileService {
     File createFile(File file);
-    File findFileById(String id);
+    Optional<File> findFileById(String id);
     List<File> getFiles();
     List<File> getFilesByUserId(String userId);
     void deleteFileById(String id);
