+++ src/main/java/fr/esgi/grp9/uparserbackend/file/web/FileController.java
@@ -1,20 +1,21 @@
 package fr.esgi.grp9.uparserbackend.file.web;
 
-import fr.esgi.grp9.uparserbackend.exception.common.NotFoundWithIdException;
 import fr.esgi.grp9.uparserbackend.file.domain.File;
-import fr.esgi.grp9.uparserbackend.file.domain.FileServiceImpl;
+import fr.esgi.grp9.uparserbackend.file.service.FileService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/files")
 public class FileController {
-    private final FileServiceImpl fileService;
+    private final FileService fileService;
 
-    public FileController(FileServiceImpl fileService) {
+    public FileController(FileService fileService) {
         this.fileService = fileService;
     }
 
@@ -30,10 +31,19 @@ public class FileController {
 
     @GetMapping("/{id}")
     public ResponseEntity<File> getFileById(@PathVariable String id){
+        if (id == null){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Empty id.");
+        }
         try {
-            return new ResponseEntity<>(fileService.findFileById(id), HttpStatus.OK);
-        } catch (NotFoundWithIdException notFoundWithIdException){
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            Optional<File> _file = fileService.findFileById(id);
+            if (_file.isEmpty()) {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "This file doesn't exist.");
+            }
+
+            return new ResponseEntity<>(_file.get(), HttpStatus.OK);
+        } catch (Exception e){
+            e.printStackTrace();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, e.getMessage());
         }
     }
 
@@ -44,23 +54,26 @@ public class FileController {
             File _file = fileService.createFile(file);
             return new ResponseEntity<>(_file, HttpStatus.CREATED);
         } catch (Exception exception){
+            exception.printStackTrace();
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteFile(@PathVariable String id) {
-
-        File _fileExist = fileService.findFileById(id);
-
-        if (_fileExist != null) {
-            try {
-                fileService.deleteFileById(id);
-                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
-            } catch (Exception exception) {
-                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
-            }
+        if (id == null){
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Empty id.");
         }
-        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        Optional<File> _file = fileService.findFileById(id);
+        if (_file.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "This file doesn't exist.");
+        }
+        try {
+            fileService.deleteFileById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, exception.getMessage());
+        }
     }
 }
